+++ packages/article-skeleton/src/article-skeleton.js
@@ -11,7 +11,9 @@ import {
   WelcomeBanner,
   ArticleSidebar,
   UpdateButtonWithDelay,
-  Banner
+  Banner,
+  SocialEmbedsProvider,
+  useSocialEmbedsContext
 } from "@times-components/ts-components";
 import { spacing } from "@times-components/ts-styleguide";
 import UserState from "@times-components/user-state";
@@ -93,6 +95,22 @@ const ArticleSkeleton = ({
   } = article;
 
   const [showVerifyEmailBanner, setShowEmailVerifyBanner] = useState(false);
+
+  const { isSocialEmbedAllowed, isAllowedOnce } = useSocialEmbedsContext();
+
+  useEffect(
+    () => {
+      // Trigger Twitter embed load when isSocialEmbedAllowed or isAllowedOnce switches to true
+      if (
+        (isSocialEmbedAllowed.twitter || isAllowedOnce.twitter) &&
+        window.twttr &&
+        window.twttr.widgets
+      ) {
+        window.twttr.widgets.load();
+      }
+    },
+    [isSocialEmbedAllowed.twitter, isAllowedOnce.twitter]
+  );
 
   const sidebarRef = useRef();
 
@@ -466,6 +484,12 @@ const ArticleSkeleton = ({
   );
 };
 
+const ArticleSkeletonWithProvider = props => (
+  <SocialEmbedsProvider>
+    <ArticleSkeleton {...props} />
+  </SocialEmbedsProvider>
+);
+
 ArticleSkeleton.propTypes = {
   ...articleSkeletonPropTypes,
   paidContentClassName: PropTypes.string
@@ -476,4 +500,6 @@ export { KeylineItem, ArticleKeylineItem } from "./keylines";
 
 export { ArticleLink };
 
-export default articleTrackingContext(withTrackScrollDepth(ArticleSkeleton));
+export default articleTrackingContext(
+  withTrackScrollDepth(ArticleSkeletonWithProvider)
+);