+++ packages/ts-components/src/components/social-embed/constants.ts
@@ -1,0 +1,14 @@
+export const modalType = {
+  GDPR: 'gdpr',
+  CCPA: 'ccpa'
+} as const;
+
+export const eventStatus = {
+  tcLoaded: 'tcloaded',
+  cmpIsShown: 'cmpuishown',
+  userActionComplete: 'useractioncomplete'
+} as const;
+
+export const vendors = {
+  twitter: 'twitter'
+} as const;