+++ packages/ts-components/src/components/social-embed/helpers/socialMediaVendors.ts
@@ -1,0 +1,15 @@
+export const socialMediaVendors: {
+  [key: string]: { id: string; status: string; title: string };
+} = {
+  twitter: {
+    id: '5fab0c31a22863611c5f8764',
+    status: 'pending',
+    title: 'X (Twitter)'
+  },
+  youtube: {
+    id: '5e7ac3fae30e7d1bc1ebf5e8',
+    status: 'pending',
+    title: 'Youtube'
+  },
+  tiktok: { id: '5e7f6927b8e05c4e491e7380', status: 'pending', title: 'Tiktok' }
+};