+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.19.28](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.27...@times-components/article-list@9.19.28) (2024-12-05)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
+## [9.19.27](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.26...@times-components/article-list@9.19.27) (2024-12-03)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.19.26](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.25...@times-components/article-list@9.19.26) (2024-12-02)
 
 **Note:** Version bump only for package @times-components/article-list