+++ packages/card/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.16.0](https://github.com/newsuk/times-components/compare/@times-components/card@6.15.2...@times-components/card@6.16.0) (2024-12-05)
+
+
+### Features
+
+* **TMC-0000:** Implement social embeds ([#3989](https://github.com/newsuk/times-components/issues/3989)) ([4a4e7cd](https://github.com/newsuk/times-components/commit/4a4e7cdceec5fe7402b4403170013f325eec8a84))
+
+
+
+
+
 ## [6.15.2](https://github.com/newsuk/times-components/compare/@times-components/card@6.15.1...@times-components/card@6.15.2) (2024-11-05)
 
 **Note:** Version bump only for package @times-components/card