+++ packages/ssr/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.59.27](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.59.26...@times-components/ssr@2.59.27) (2024-12-05)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.59.26](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.59.25...@times-components/ssr@2.59.26) (2024-12-03)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
 ## [2.59.25](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.59.24...@times-components/ssr@2.59.25) (2024-12-02)
 
 **Note:** Version bump only for package @times-components/ssr